+++ dune/dpg/functions/gridviewfunctions.hh
@@ -6,6 +6,7 @@
 #include <type_traits>
 #include <utility>
 
+#include <dune/common/fvector.hh>
 #include <dune/common/std/type_traits.hh>
 #include <dune/common/typeutilities.hh>
 #include <dune/functions/gridfunctions/gridviewentityset.hh>